-  refactor messaging app views and serializers  -  write unit tests for list dm messages view
