+++ quexl/apps/messaging/renderers.py
@@ -1,0 +1,1 @@
+from quexl.utils.renderers import DefaultRenderer as MessagesRenderer  # noqa
